+++ offer-service/src/main/java/com/github/bovvver/offermanagment/offercreation/CreateOfferCommand.java
@@ -2,12 +2,13 @@ package com.github.bovvver.offermanagment.offercreation;
 
 import com.github.bovvver.offermanagment.vo.Location;
 
+import java.math.BigDecimal;
 import java.util.Set;
 
 public record CreateOfferCommand (
         String title,
         String description,
-        double salary,
+        BigDecimal salary,
         Location location,
         Set<String> serviceCategories
 ) {
